remove unused components and folders and move LocationListContainer into a folder
