+++ src/Components/App/App.js
@@ -3,7 +3,7 @@ import './App.css';
 import Header from '../Header/Header.js';
 import { setLocations } from '../../actions/index.js';
 import { connect } from 'react-redux';
-import LocationListContainer from '../LocationListContainer';
+import LocationListContainer from '../LocationListContainer/LocationListContainer.js';
 
 
 class App extends Component {
@@ -15,6 +15,7 @@ class App extends Component {
   }
 
   componentDidMount () {
+
   }
 
   render() {
