--- src/Components/MapContainer/MapContainer.js
@@ -1,1 +1,0 @@
-import Map from '../Map/Map.js';
